refactor: extract shared item widgets and remove duplicate code
- Merged LostItemData/FoundItemData into a single ItemData class (with isLost field)
- Extracted ItemSearchBar, ItemCategoryChip, ItemEmptyState, ItemCard, ItemDetailsView into item_widgets.dart
- Deleted lost_item_details_view.dart and found_item_details_view.dart (consolidated into ItemDetailsView)
- Deleted empty placeholder files: custom_button.dart, custom_text_field.dart
- Reduced lost_items_view and found_items_view from ~550 lines each to ~170 lines each (~1200 lines removed)
