--- lost_found_app/lib/views/lost/lost_item_details_view.dart
@@ -1,301 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../../utils/app_theme.dart';
-import 'lost_items_view.dart';
-
-class LostItemDetailsView extends StatelessWidget {
-  const LostItemDetailsView({
-    super.key,
-    required this.item,
-    this.scrollController,
-  });
-
-  final LostItemData item;
-  final ScrollController? scrollController;
-
-  @override
-  Widget build(BuildContext context) {
-    final textTheme = Theme.of(context).textTheme;
-
-    return Container(
-      width: double.infinity,
-      decoration: const BoxDecoration(
-        color: AppTheme.surface,
-        borderRadius: BorderRadius.vertical(top: Radius.circular(40)),
-      ),
-      child: SingleChildScrollView(
-        controller: scrollController,
-        padding: EdgeInsets.only(
-          bottom: MediaQuery.of(context).padding.bottom + 12,
-        ),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Center(
-              child: Container(
-                width: 54,
-                height: 6,
-                margin: const EdgeInsets.only(top: 12),
-                decoration: BoxDecoration(
-                  color: AppTheme.border,
-                  borderRadius: BorderRadius.circular(999),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(26, 18, 22, 0),
-              child: Row(
-                children: [
-                  Expanded(
-                    child: Text(
-                      'Item Details',
-                      style: textTheme.headlineMedium?.copyWith(
-                        color: AppTheme.secondary,
-                        fontSize: 24,
-                      ),
-                    ),
-                  ),
-                  IconButton(
-                    onPressed: () => Navigator.of(context).pop(),
-                    icon: const Icon(Icons.close_rounded, size: 30),
-                    color: AppTheme.secondary,
-                  ),
-                ],
-              ),
-            ),
-            const SizedBox(height: 8),
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 26),
-              child: ClipRRect(
-                borderRadius: BorderRadius.circular(28),
-                child: Container(
-                  height: 160,
-                  width: double.infinity,
-                  decoration: BoxDecoration(
-                    gradient: LinearGradient(
-                      colors: [item.backgroundColor, item.accentColor],
-                      begin: Alignment.topLeft,
-                      end: Alignment.bottomRight,
-                    ),
-                  ),
-                  child: Align(
-                    alignment: Alignment.bottomLeft,
-                    child: Padding(
-                      padding: const EdgeInsets.all(20),
-                      child: Icon(
-                        item.icon,
-                        size: 120,
-                        color: Colors.white.withValues(alpha: 0.9),
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(26, 26, 26, 0),
-              child: Row(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Expanded(
-                    child: Text(
-                      item.title,
-                      style: textTheme.headlineLarge?.copyWith(
-                        color: AppTheme.secondary,
-                        fontSize: 30,
-                        height: 1.15,
-                      ),
-                    ),
-                  ),
-                  const SizedBox(width: 14),
-                  const _StatusBadge(label: 'Lost'),
-                ],
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(26, 18, 26, 0),
-              child: Text(
-                item.description,
-                style: textTheme.bodyLarge?.copyWith(
-                  color: AppTheme.textSecondary,
-                  fontSize: 17,
-                  height: 1.45,
-                ),
-              ),
-            ),
-            const SizedBox(height: 20),
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 26),
-              child: Column(
-                children: [
-                  _DetailInfoCard(
-                    icon: Icons.sell_outlined,
-                    label: 'Category',
-                    value: item.category,
-                  ),
-                  const SizedBox(height: 12),
-                  _DetailInfoCard(
-                    icon: Icons.location_on_outlined,
-                    label: 'Location',
-                    value: item.location,
-                  ),
-                  const SizedBox(height: 12),
-                  _DetailInfoCard(
-                    icon: Icons.access_time_outlined,
-                    label: 'Date',
-                    value: item.date,
-                  ),
-                ],
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(26, 24, 26, 0),
-              child: Divider(
-                color: AppTheme.border.withValues(alpha: 0.85),
-                height: 1,
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(26, 26, 26, 0),
-              child: Text(
-                'Contact Information',
-                style: textTheme.headlineMedium?.copyWith(
-                  color: AppTheme.secondary,
-                  fontSize: 22,
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(26, 16, 26, 0),
-              child: Container(
-                width: double.infinity,
-                padding: const EdgeInsets.all(24),
-                decoration: BoxDecoration(
-                  gradient: const LinearGradient(
-                    colors: [Color(0xFFEAF2FF), Color(0xFFD9E8FF)],
-                    begin: Alignment.topLeft,
-                    end: Alignment.bottomRight,
-                  ),
-                  borderRadius: BorderRadius.circular(24),
-                ),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      'If you found this item, please contact:',
-                      style: textTheme.bodyLarge?.copyWith(
-                        color: AppTheme.textSecondary,
-                        fontSize: 16,
-                      ),
-                    ),
-                    const SizedBox(height: 16),
-                    Text(
-                      item.contactEmail,
-                      style: textTheme.headlineMedium?.copyWith(
-                        color: AppTheme.primary,
-                        fontSize: 22,
-                        height: 1.2,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(26, 22, 26, 0),
-              child: SizedBox(
-                width: double.infinity,
-                child: ElevatedButton(
-                  onPressed: () => Navigator.of(context).pop(),
-                  child: const Text('Close'),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _StatusBadge extends StatelessWidget {
-  const _StatusBadge({required this.label});
-
-  final String label;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 22, vertical: 16),
-      decoration: BoxDecoration(
-        color: const Color(0xFFFDE2E2),
-        borderRadius: BorderRadius.circular(999),
-      ),
-      child: Text(
-        label,
-        style: Theme.of(context).textTheme.titleMedium?.copyWith(
-          color: const Color(0xFFDC2626),
-          fontWeight: FontWeight.w700,
-        ),
-      ),
-    );
-  }
-}
-
-class _DetailInfoCard extends StatelessWidget {
-  const _DetailInfoCard({
-    required this.icon,
-    required this.label,
-    required this.value,
-  });
-
-  final IconData icon;
-  final String label;
-  final String value;
-
-  @override
-  Widget build(BuildContext context) {
-    final textTheme = Theme.of(context).textTheme;
-
-    return Container(
-      width: double.infinity,
-      padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 18),
-      decoration: BoxDecoration(
-        color: const Color(0xFFF8FAFD),
-        borderRadius: BorderRadius.circular(22),
-      ),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Padding(
-            padding: const EdgeInsets.only(top: 2),
-            child: Icon(icon, color: AppTheme.primary, size: 30),
-          ),
-          const SizedBox(width: 18),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  label,
-                  style: textTheme.bodyLarge?.copyWith(
-                    color: AppTheme.textSecondary,
-                    fontSize: 15,
-                  ),
-                ),
-                const SizedBox(height: 6),
-                Text(
-                  value,
-                  style: textTheme.headlineMedium?.copyWith(
-                    color: AppTheme.secondary,
-                    fontSize: 19,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
